Deduplicate requestAnimationFrame scheduling code
This unifies handling of array vs. object in the `data` prop. Code for
setting up the animation was duplicated, and really they're doing nearly
the same thing: if `data` is an array, extend the animation queue, if
it's an object, then replace the animation queue with `data`.

It also fixes a couple things:
* If the initial `data` value is an array, then upon mount we'll start
  animating, starting at the first value, instead of ignoring everything
  but the first value.
* Keep a handle to the `setTimeout` that controls our delay, because
  we'll need to cancel it when new data comes in or we unmount.

/cc @kenwheeler @boygirl