Use d3-timer instead of manual requestAnimationFrame
Different from Guy's PR: this is based on the merge of our dedupe, and
fixes the case where we have multiple items in the queue and/or have an
`onEnd`.

Totally unrelated change, but might as well get it in: use `breakConfig`
in `.babelrc`.

/cc @boygirl @guykr-stratoscale

TODO: Add Guy to a `CONTRIBUTORS` file/section somewhere? We have none
at the moment.