Add support for providing label text as a prop
This is a fall back in the absence of children for the label. Custom label
components will receive label text through this prop when created by
victory components. Custom label components will not have their children
modified. Therefore, if we use VictoryLabel as a custom label component
as in:
https://github.com/FormidableLabs/victory-chart/blob/c72cfe78cb4a6091fa2a77e730bbce7ef78ba16e/demo/components/victory-scatter-demo.jsx#L101
then we need to allow for setting the label text through props.