Moves label text to props rather than children
Depends on FormidableLabs/victory-core/#10 We were previously assuming
that the children of custom labels contained accurate label information.
However, as discussed in FormidableLabs/victory-chart/#101 we are moving
to an api that supports making many clones of a single custom label
component based on an array of label values. That use case is clear in
not assuming that the children of the label 'factory' already know the
label text, it can't, we want to explicitly provide it.

So, I am proposing we eschew handling of injected components children.
Instead, we will provide our best guess at label text as a prop.

Match feedback in FormidableLabs/victory-core#10