Adds dataComponent prop to Line
Includes a (single) shallow test to assert the type of node matches
that provided in the props. I noticed we have a set of tests designed to
test the segmentation of datasets. I chose to add this test using the
simplest case.

Missing: an example in the victory line demos! I am thinking something
that takes a size and draws line segments with different stroke widths.

Contributes to resolution of FormidableLabs/victory-chart#100