Pie Label Positioning
Added the option for "startAngle" and "endAngle" for the label to be
positioned.

The default is "centroid" which does not altere the original behavior.

As posted on the issue #1117 it would be nice to position the label
inside the actual slice but as per one rule of no-magic-numbers I can't
multiply by 0.9 to do this equally for all labels. This could be left
for later.

Closes #1117