Add victory tooltip story and fix label bug
It turns out that the y-position of the label component is calculated
using the inline font size. If there is no font size, the y-position
will evaluate to NaN. For now, this fix continues to return the fontSize
style attribute and use it as an inline style. If we want to calculate
the size dynamically, that might be a bigger fix later.