Add yarn version config
This ensures that yarn commands in the Victory repo will default to yarn
v1 regardless of the specific version that is installed on your machine.
Yarn v2 is not compatible with lerna, so we will need to leave our yarn
version pinned to v1 until we swap out our monorepo tooling.