Remove packages from chromatic so new demo stories are not getting UI tested
For now, we should only run stories in /stories in the chromatic UI
tests. Any stories we are adding in /packages are just WIP demos to help
with refactoring, and they don't need to be tested in chromatic. If we
co-locate stories in victory packages, we need to revert this change.